fix: build 오류 수정
