+++ src/pages/_document.tsx
@@ -1,4 +1,5 @@
 import { Head, Html, Main, NextScript } from "next/document"
+import Script from "next/script"
 
 function Document() {
   return (
@@ -9,11 +10,11 @@ function Document() {
           as="style"
           href="https://cdn.jsdelivr.net/gh/orioncactus/pretendard/dist/web/static/pretendard-dynamic-subset.css"
         />
-        <script
+        <Script
           src="https://t1.kakaocdn.net/kakao_js_sdk/2.1.0/kakao.js"
           integrity="sha384-OfbOqPoV2XcfZpqrLgqYCNSNBJW4JU/lLrtKk0cpkWvCrDRotHaQ9SSMGeP7u8NB"
           crossOrigin="anonymous"
-        ></script>
+        ></Script>
       </Head>
       <body>
         <Main />
